Added brain, team module, misc.
The brain of the gladiators has nearly been hooked up, but there is still a lot
more work to do. The next steps will be detecting collisions and making the
projectiles interact with the gladiators.

There have been lots of changes to the rest of the module as well, mostly to
accommodate the brain and it's input/output.