Mutation added, fitness selection, parameters adjusted, ...
A box is now drawn around the arena.

Copy for gladiators and brains is now implemented, used in the mutation and
fitness functions which are now present and active. They do not work very well,
but that is because of a combination of bugs and incorrect logic present in the
program, most of the functionality has been added, it just needs testing.

Next steps:
* Solve bugs in collision detection
* General improvements
* Tweak algorithms