Added limits to turn rate, tweaked config
The turn rate is now limited, which may or may not help things. The collision
detection is still the main problem.

Perhaps the arena is best not being wrapped around as well.

The Read-Me file has been updated to reflect all the changes.

The configuration file has also been tweaked.