Added limits, improved brain, fitness, new inputs, command line options, ...
The brain now has a better layer mechanism, it will still have to be improved,
but it is more extensible. The brain in now separated out into layer_t
structures.

The number of gladiators is now configurable.

The arena now can use limits instead of wrapping when an object hits the edge
of the arena. Instead of wrapping around an object can get stuck at the limit.

There are new inputs to a gladiator, including a state variable, random input,
wall detection and enemy collision.

The output for the field of view has been improved.

The fitness function is now both configurable and has more options.

There are now command line options that can be passed to the program.