Fixed collision detection, many other changes.
It is now possible to single step the simulation, by pressing 's' a single tick
of the world will be executed, and then the simulation will be paused again.

More information can be printed onto the screen, such as mutation rates and
parents fitness levels.

Many variables have been renamed so they better fit a naming convention.

Some functionality has been moved around into vars (anything relating to log
levels and debug printing is now in "vars.c").

The collision detection has been fixed, so the inputs to the gladiators are
finally correct, the debug information it draws the screen when a collision is
detected (if set by the log level) is now also correct.

A message is printed to the screen when the simulation is paused.

A bug with the LDFLAGS variable in the makefile has been fixed, which was only
triggered on some machines, it was caused by specifying the libraries before
the file they are used in on the command line, a classic mistake.

The random number generator is now centralized into the "util.c" file.

There are now more log levels as well, and they are used.