Added headless mode, some assertions, doxygen headers, moved collision functions, ...
A headless mode has been added which runs quicker than the GUI program
(although it still a lot slower than it could be, mainly due to a slow
collision detection algorithm which could be improved on).

Assertions have been added to most functions, however there are still many that
need to be added, most of the assertions deal with NULL pointers, more that
deal with the program logic should be added.

More @todo notes have been added, some will now need to be removed.

Doxygen headers, including a license (which will need to be changed to
something more sensible), have been prepended to each file.