Added food, world object, misc.
Food has been added, which gladiators can consume for energy and fitness
(although not for health at the moment).

A world object has been added which should, but does not at the moment, contain
the entire state of the program.

There have been many other minor changes and improvements as well.