Changed licenses, added s-expression parser, split GUI code
License has been changed to the MIT license.

An S-expression parser has been added, this will be used for storing world
state and configuration information moving forward. Some way of validating it
and making a menu automatically from it could be in the long term.

The GUI code has been split out of the util.c file, into its own.