Documentation, serialization functionality, ...
The documentation has been improved a lot, but there is more to do in
describing how the program works. All of the configuration items so far have
now been documented. An image of the game play has also been added.

The makefile now includes rules to build the documentation.

The serialization/deserialization functionality has been improved, the method
by which objects are serialized still needs a lot of improvement however and I
am sure there will be a lot of bug fixing to do.