Vastly improved speed of game, improved S-Expression utils, ...
The speed of the game has been improved a lot by improving the collision
detection algorithm (by making it a lot more sensible than it was before).

The S-Expression now has a printf like function for generating S-Expressions,
it needs testing however, this will be used for serialization.

The defaults for some of the variables have changes again.