Mutation mechanism, defaults, projectile mechanics
A default configuration file has been added, and the defaults changed in the
program. A larger population in a larger arena seems to produce better results.

The mutation mechanism has been reworked to be less drastic, it was made this
way by mistake previously, and has now been tamed.

The projectile mechanics has been changed, so a pool of free projectile objects
can be fire from, allowing multiple projectiles to be fired by the same
gladiator. With the wrong defaults this allows a single gladiator to spam the
arena with projectiles and exhaust the free projectile pool. This should be
corrected with a max number of projectiles per gladiator rule.