Moved help, serialization, state option, more
* An option for auto-incrementing the field of view has been added, it
defaults to off.
* An option to add a state variable to each neuron has been added, it
defaults to off.
* An option for printing some statistics about a gladiator next to the
gladiator has been added.
* Some of the defaults for printing have been changed.
* The help section has been removed from the 'readme.md' file to where
the variables are defined in order to keep them in sync more.
* The state of the interpreter can, seemingly correctly, be serialized
and deserialized allowing the saving of good runs.
* More comments about the direction of the interpreter have been added.

Of note:
* Valgrind shows copious amounts of errors, they all seem to be related
to the OpenGl/Glut stack however, it makes it difficult to find out if
there is actually a problem in my code or not.
* The state-variable stuff does not seem to work too well.