(broken) Serialization bugs; debug commit
There appears to be serious problems with the serialization code, which
is a shame, that would allow the loading and saving on previously
evolved gladiators, perhaps allowing the production of better models.

Some code changes have been made to attempt to debug that, but nothing
is forthcoming at the moment.

Some more TODO comments have been added as well.